requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pandas-market-calendars>=4.3.1
pyarrow>=14.0.0
python-binance>=1.0.19
numba>=0.58.0
structlog>=24.1.0
python-telegram-bot>=21.0.1
matplotlib>=3.7.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Development & testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
ruff>=0.1.0
mypy>=1.5.0
pre-commit>=3.5.0
